Sensorlib-v2: Move all the code from sensorlib inside `v2` folder
This commit moves all the code of the previous implementation of sensorlib inside the `v2` folder.
