+++ mltest/src/main/java/com/handmonitor/mltest/service/TestService.kt
@@ -4,8 +4,8 @@ import android.app.Service
 import android.content.Intent
 import android.os.IBinder
 import android.util.Log
-import com.handmonitor.sensorlib.SensorWindowProducer
-import com.handmonitor.sensorlib.asFlow
+import com.handmonitor.sensorlib.v2.SensorWindowProducer
+import com.handmonitor.sensorlib.v2.asFlow
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
