+++ sensorlib/src/main/java/com/handmonitor/sensorlib/v2/SensorSampleFilter.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorlib
+package com.handmonitor.sensorlib.v2
 
 import android.hardware.SensorEvent
 import android.util.Log
