+++ sensorlib/src/main/java/com/handmonitor/sensorlib/v2/SensorWindow.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorlib
+package com.handmonitor.sensorlib.v2
 
 import java.nio.FloatBuffer
