+++ sensorlib/src/main/java/com/handmonitor/sensorlib/v2/SensorWindowFlow.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorlib
+package com.handmonitor.sensorlib.v2
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.awaitClose
