+++ sensorlib/src/test/java/com/handmonitor/sensorlib/v2/SensorWindowBufferTest.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorlib
+package com.handmonitor.sensorlib.v2
 
 import com.google.common.truth.Truth.assertThat
 import io.mockk.junit5.MockKExtension
