+++ sensorlib/src/test/java/com/handmonitor/sensorlib/v2/SensorWindowFlowTest.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorlib
+package com.handmonitor.sensorlib.v2
 
 import com.google.common.truth.Truth.assertThat
 import io.mockk.Runs
