+++ sensorlib/src/test/java/com/handmonitor/sensorlib/v2/SensorWindowTest.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.sensorlib
+package com.handmonitor.sensorlib.v2
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.jupiter.api.Test
